backend/requirements.txt
sanic[ext]
SQLAlchemy==1.4.44
PyMySQL[rsa]
PyMySQL==1.0.2
python-dotenv==0.21.0
bcrypt==4.0.1
pyjwt==2.6.0
pytz==2021.1
apscheduler==3.9.1
pandas==1.5.3
openpyxl==3.1.1